ARM: cns3xxx: fix mapping of private memory region
Since commit 0536bdf33faf (ARM: move iotable mappings within the vmalloc
region), the Cavium CNS3xxx cannot boot anymore.

This is caused by the pre-defined iotable mappings is not in the vmalloc region.
This patch move the iotable mappings into the vmalloc region, and merge the
MPCore private memory region (containing the SCU, the GIC and the TWD) as a
single region.

Signed-off-by: Mac Lin <mkl0301@gmail.com>